updated repositories

+++ drone-implementation/drone-DataAccess/Repositories/GenericRepository.cs
@@ -34,7 +34,7 @@ namespace drone_DataAccess.Repositories
             return _context.Set<T>().ToList();
         }
 
-        public T GetById(long id)
+        public T GetById(int id)
         {
             return _context.Set<T>().Find(id);
         }
